+++ games/SparkleDie/SparkleDie.styles.ts
@@ -3,7 +3,6 @@ import css from 'styled-jsx/css';
 export const getContainerCSS = (fontFamily: string) => css.resolve`
    {
     font-family: ${fontFamily};
-    background-color: #fff;
     overflow: hidden;
     height: 100vh;
     display: flex;