+++ CHANGELOG.md
@@ -6,13 +6,15 @@ All notable changes, updated and additions to this site will be documented below
 
 **Added**
 
-- **New Game** - [First and Last Letter](https://www.eslintherok.com/game/first_and_last_letter)
+- New Game - [First and Last Letter](https://www.eslintherok.com/game/first_and_last_letter)
 - An animated dot on the nav changelog link to indicate an update has occured and hasn't been reviewed
 - The date a game is published, so we can show a _New_ badge on the games page
 
 **Changed**
 
 - Updated [Privacy Policy](https://www.eslintherok.com/privacy)
+- Updated Game - [Sparkle Die](https://www.eslintherok.com/game/sparkle_die)
+  - Sparkle Die now uses both vocabulary and expressions per user request
 
 ## [2.0.0] - 2021-10-04